refactoring for types
