+++ src/types/index.ts
@@ -1,7 +1,21 @@
+import { Intent, Knot, Story, TrainingMaterial } from "../generated/client/models";
+
 /**
  * Interface describing an access token
  */
 export interface AccessToken {
   token: string;
   userId: string;
+};
+
+/**
+ * Story data
+ */
+export interface StoryData {
+  story?: Story;
+  knots?: Knot[];
+  intents?: Intent[];
+  selectedKnot?: Knot;
+  selectedIntent? : Intent;
+  trainingMaterial?: TrainingMaterial[];
 };
